Marks changes

+++ update.sh
@@ -1,9 +1,7 @@
 #!/bin/sh
 
-# The following line is important to ensure this script stops if this repo or submodules have oustanding changes that would prevent a git pull.
+# The following line is important to ensure this script stops if any repos have outstanding changes that would prevent a git pull.
 set -e
-
-set -x
 
 getUpdateScriptInfo () {
   ls -l update.sh
@@ -11,8 +9,15 @@ getUpdateScriptInfo () {
 
 oldUpdateScriptInfo=`getUpdateScriptInfo`
 
-git checkout main > /dev/null
+echo Updating redcap_cypress_docker...
 git pull
+git fetch https://github.com/vanderbilt-redcap/redcap_cypress_docker main
+commitsBehindMain=`git log --oneline ..FETCH_HEAD | wc -l`
+if [ $commitsBehindMain != 0 ]; then
+    echo
+    echo Please either checkout the main branch for redcap_cypress_docker, or merge it into your working branch.
+    exit
+fi
 
 if [ "$oldUpdateScriptInfo" != "`getUpdateScriptInfo`" ]; then
     echo A changed to update.sh was detected.  Restarting this script...
@@ -20,6 +25,7 @@ if [ "$oldUpdateScriptInfo" != "`getUpdateScriptInfo`" ]; then
     exit
 fi
 
+echo Updating redcap_rsvc...
 cd redcap_cypress/redcap_rsvc
 git fetch https://github.com/vanderbilt-redcap/redcap_rsvc staging
 rsvcBranchName=`git rev-parse --abbrev-ref HEAD`
@@ -31,6 +37,7 @@ if [ "$rsvcBranchName" = "staging" ] || [ "$rsvcBranchName" = "main" ]; then
 else
     commitsBehindStaging=`git log --oneline ..FETCH_HEAD | wc -l`
     if [ $commitsBehindStaging != 0 ]; then
+        set +x
         echo
         echo Please merge the latest from the 'staging' branch into your redcap_rsvc branch.
         echo This is not performed automatically to avoid interfering with any active development. 
@@ -39,17 +46,35 @@ else
 fi
 cd ../..
 
+echo Updating redcap_cypress...
 cd redcap_cypress
-git checkout master
 git pull
+git fetch https://github.com/vanderbilt-redcap/redcap_cypress master
+commitsBehindMaster=`git log --oneline ..FETCH_HEAD | wc -l`
+if [ $commitsBehindMaster != 0 ]; then
+    echo
+    echo Please either checkout the master branch for redcap_cypress, or merge it into your working branch.
+    exit
+fi
+
 # Adam Lewis had an instance where cypress started throwing confusing errors on every feature which was resolved by the following steps:
 rm node_modules -r
 npm cache clean --force
 npm install --no-fund --no-audit
 cd ..
 
+echo Updating redcap_docker...
 cd redcap_docker
-git checkout main
 git pull
+git fetch https://github.com/vanderbilt-redcap/redcap_docker main
+commitsBehindMain=`git log --oneline ..FETCH_HEAD | wc -l`
+if [ $commitsBehindMain != 0 ]; then
+    echo
+    echo Please either checkout the main branch for redcap_docker, or merge it into your working branch.
+    exit
+fi
 docker compose --profile external-storage --profile sftp down # This ensures a running container is restarted, which can fix various docker issues.
 docker compose up -d --build --remove-orphans # This ensures the container is rebuilt to include any Dockerfile changes, other updates, or fix various issues.
+
+echo
+echo 'Update completed successfully!'
